pyproject.toml
[project]
name = "industriel-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "matplotlib>=3.10.8",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
